feat: split the codec and server into their own independent modules.
