+++ codec/testdata/TestV1Codec/payload_exceeds_max_byte_limit
