+++ codec/testdata/TestV2Codec/payload exceeds max byte limit
