+++ server/cmd/main_test.go
@@ -6,15 +6,14 @@ package main
 
 import (
 	"context"
-	"os"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
 	"github.com/DataDog/temporal-large-payload-codec/server/storage"
 	"github.com/DataDog/temporal-large-payload-codec/server/storage/gcs"
 	"github.com/DataDog/temporal-large-payload-codec/server/storage/memory"
 	"github.com/DataDog/temporal-large-payload-codec/server/storage/s3"
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
 const (
