+++ server/e2e_test.go
@@ -7,6 +7,10 @@ package server_test
 import (
 	"context"
 	"fmt"
+	"github.com/DataDog/temporal-large-payload-codec/codec"
+	"github.com/DataDog/temporal-large-payload-codec/server"
+	"github.com/DataDog/temporal-large-payload-codec/server/storage/memory"
+	"go.temporal.io/api/enums/v1"
 	"math/rand"
 	"net/http/httptest"
 	"testing"
@@ -14,15 +18,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"github.com/temporalio/temporalite/temporaltest"
-	"go.temporal.io/api/enums/v1"
 	"go.temporal.io/sdk/client"
 	"go.temporal.io/sdk/converter"
 	"go.temporal.io/sdk/worker"
 	"go.temporal.io/sdk/workflow"
-
-	codec "github.com/DataDog/temporal-large-payload-codec"
-	"github.com/DataDog/temporal-large-payload-codec/server"
-	"github.com/DataDog/temporal-large-payload-codec/server/storage/memory"
 )
 
 const (
