+++ server/logging/stdlogger.go
