+++ server/storage/azure/azure.go
@@ -7,13 +7,12 @@ package azure
 import (
 	"context"
 	"fmt"
+	"github.com/DataDog/temporal-large-payload-codec/server/storage"
 	"io"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/bloberror"
-
-	"github.com/DataDog/temporal-large-payload-codec/server/storage"
 )
 
 // Config provides all configuration to create the Azure based driver for LPS.
