+++ server/storage/s3/s3.go
@@ -8,15 +8,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/DataDog/temporal-large-payload-codec/server/storage"
+	"github.com/aws/smithy-go"
 	"io"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	s3types "github.com/aws/aws-sdk-go-v2/service/s3/types"
-	"github.com/aws/smithy-go"
-
-	"github.com/DataDog/temporal-large-payload-codec/server/storage"
 )
 
 // Config provides all configuration to create the S3 based driver for LPS.
